change from % commission to fixed amount fix normalizer, add calculations of very tiny precision

+++ src/Contracts/FeeCalculatorInterface.php
@@ -1,11 +1,13 @@
-<?php declare(strict_types=1);
+<?php
+
+declare(strict_types=1);
 
 namespace Contracts;
 
 interface FeeCalculatorInterface
 {
     /**
-     * @return int The calculated total fee.
+     * @return int the calculated total fee
      */
-    public function calculate(int $term, float $amount): int;
+    public function calculate(int $term, float $amount): float;
 }
